Refactor: moved DataPropertyEmitter to DataNodeEmitter file
